5.1th Minor fix

+++ FirstYearExamination/Tower/Tower_stats/Slow_Tower.cs
@@ -11,7 +11,7 @@ namespace FirstYearExamination
 {
     public class Slow_Tower : New_Tower
     {
-        public Slow_Tower(int damage, int range, int fireRate, int projectileSpeed, int cost, string name, ProjectileType projectileType)
+        public Slow_Tower(int damage, int range, float fireRate, int projectileSpeed, int cost, string name, ProjectileType projectileType)
         {
             this.damage = damage;
             this.range = range;
